correccion getPlayer del map,si no esta devuelve nullptr

+++ src/Servidor/Modelo/Character.cpp
@@ -68,7 +68,7 @@ float Character::calculateRecoverMana(float seconds) const {
     return (float)recoveryFactor * seconds;
 }
 
-int Character::calculateRecoverManaMeditating(int seconds) const {
+float Character::calculateRecoverManaMeditating(float seconds) const {
     return meditationRecoveryFactor * seconds;
 }
